delete source refs to public folder
index.html is in the public folder.  The references to all the library and js code folders is wrong.  The 'Public' folder is assumed because the index.html file is being loaded from the 'Public' folder.  Index.html could be put in the DrEdit root directory, but there is an index.php file there.  Maybe not a good idea to have them together.